### 2023-10-27 #5992 @editorialbot
[PRE REVIEW]: aion: An R Package to Represent Archaeological Time Series
**Submitting author:** @nfrerebeau (<a href="http://orcid.org/0000-0001-5759-4944">Nicolas Frerebeau</a>)
**Repository:** https://github.com/tesselle/aion
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
