### 2023-10-29 #5992 @martinfleis
Hi @nfrerebeau,

I'll be editing your submission. Before we move forward, I have initiated a scope check, to ensure the package fits JOSS's [Substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). The package has a borderline size and limited functionality, so I want to make are the editorial board considers it within scope before inviting reviewers.This will take roughly two weeks.

You are invited to provide further explanation helping us to make a decision.

Thanks!
