### 2023-11-20 7dff9a4
Update pkgdown configuration
