+++ pkgdown/_pkgdown.yml
@@ -2,9 +2,6 @@ url: https://packages.tesselle.org/aion
 
 template:
   package: tekton
-  opengraph:
-    twitter:
-      creator: "@nfrerebeau"
 
 reference:
 - title: Calendars
