### 2023-11-27 #5992 @nfrerebeau
@martinfleis I was waiting for feedback from the scope check to know more precisely what additional information to provide. I just realized that I misunderstood your expectations, and that this information is necessary *for* the scope check :sweat_smile: . I'll provide you with further explanations over the course of the week.
