### 2023-11-28 #5992 @nfrerebeau
Considering the JOSS criteria, it seems to me that `aion` is a significant contribution as it offers a simple, easy-to-use solution to a common problem.

The aim of this package is to make it easier to work with time series in archaeology: it fills a gap, as there is currently no package specifically designed for this kind of series (ie. involving dates very far in the past and using different calendars). It is common practice in archaeology to use and mix data expressed in different calendars (or eras). `aion` offers a consistent framework by expressing internaly all dates in *rata die*, thus allowing arithmetic operations and comparisons, whatever the calendars used initially.

`aion` provides basic tools, and its functionality can easily be reused and extended. It is used by [`kairos`](https://github.com/tesselle/kairos) and two other projects currently under development: [`ananke`](https://github.com/tesselle/ananke) and v2.0 of [`ArchaeoPhases`](https://github.com/ArchaeoStat/ArchaeoPhases).
