### 2023-11-28 #5992 @martinfleis
@nfrerebeau Thanks! I hope we'll wrap the scope query soon.
