### 2023-12-04 d7beae6
Replace rmarkdown with markdown for generating vignettes
