+++ DESCRIPTION
@@ -31,7 +31,7 @@ Imports:
     methods
 Suggests: 
     knitr,
-    rmarkdown,
+    markdown,
     rsvg,
     svglite,
     tinysnapshot,
