### 2023-12-18 #5992 @martinfleis
@nfrerebeau Apologies it took so long but the editorial board considers the submission in scope.

if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
