### 2023-12-29 #5992 @nfrerebeau
[steko](https://reviewers.joss.theoj.org/reviewers/1821), [benmarwick](https://reviewers.joss.theoj.org/reviewers/41) and [sebastien-plutniak](https://reviewers.joss.theoj.org/reviewers/1968) are potential reviewers.
