### 2024-01-03 #5992 @sebastien-plutniak
> Hi @sebastien-plutniak and @benmarwick,
>
> would you be willing to review this submission for JOSS? It seems to be very much in your area of expertise and the codebase is relatively small, so it should not take a long time either. Thanks!

Hi @martinfleis , sure, I can review this submission!
