### 2024-01-10 #5992 @martinfleis
Hi @steko, would you be willing to eview this submission for JOSS? It seems to be very much in your area of expertise and the codebase is relatively small, so it should not take a long time either. Thanks!
