### 2024-01-12 tesselle/aion#2 @steko
JOSS review: Contribution and authorship
Hello, I am one of the reviewers for the JOSS submission described at https://github.com/openjournals/joss-reviews/issues/6210 👋

As I go through the checklist, I find this discrepancy:
- in the paper, the author is: Nicolas Frerebeau
- in the documentation (e.g. at https://packages.tesselle.org/aion/authors.html) the authors are listed as: Nicolas Frerebeau and Joe Roe, with Brice Lebrun as "Contributor"

Could you please clarify the authorship, either in the paper or in the documentation? Thanks!
