### 2024-01-12 tesselle/aion#2 @nfrerebeau
Brice Lebrun is credited as a contributor, as he designed the package logo (I updated the DESCRIPTION file).

Joe Roe is credited as author, as his 'era' package gave me the inspiration for 'aion' (as stated in the paper's acknowledgements). It seems fair to me to credit him in this way. I suggested to Joe Roe that he co-sign the paper, but he declined, considering it not justified (maybe @joeroe can confirm?).
