### 2024-01-12 tesselle/aion#2 @joeroe
Can confirm, I'm very grateful for the acknowledgement but I think all credit for the work on this package goes to Nicolas :)
