### 2024-01-15 tesselle/aion#2 @steko
Ok, I don't think it makes sense to split hairs on this. Maybe the short paragraph in the Acknowlegements section of the paper could be made more explicit?
