### 2024-01-15 903666b
Make the Acknowlegements section more explicit (#2)
