+++ DESCRIPTION
@@ -25,7 +25,7 @@ BugReports: https://github.com/tesselle/aion/issues
 Depends: 
     R (>= 3.3)
 Imports: 
-    arkhe (>= 1.4.0),
+    arkhe (>= 1.5.0),
     graphics,
     grDevices,
     methods
@@ -41,7 +41,7 @@ VignetteBuilder:
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.0
 Collate: 
     'AllClasses.R'
     'AllGenerics.R'
