### 2024-02-21 #6210 @sebastien-plutniak
Disclaimer: although the author and I do not directly collaborate on scientific products (leading projects, co-authoring papers, etc.) we both are part of the small community of archeologists interested in computer applications in France and have contact in this context.  This fact does not seem to conflict with JOSS COI policy.

After reviewing it, this submission seems almost ready for publication in JOSS. The paper and the package documentation are clear, well structured, and including a sufficient number of examples. Small improvements might be suggested, including:
* adding automated tests
* adding community guidelines (note that a code of conduct is already present)
Regarding the quality of writing, it looks correct to me, but English is not my primary language so I cannot  judge it.
