### 2024-02-26 tesselle/aion#4 @nfrerebeau
Add interoperability examples
Add examples of how `aion` can make interoperable formats used by other packages, as suggested by @sebastien-plutniak in https://github.com/openjournals/joss-reviews/issues/6210.
