### 2024-02-26 #6210 @nfrerebeau
>     * adding automated tests

There are automated tests, but `aion` uses [`tinytest`](https://github.com/markvanderloo/tinytest) instead of `testthat`. Tests can be run with `tinytest::run_test_dir()` or `devtools::check()` (`devtools::test()` won't work, as it relies on `testthat`).

>     * adding community guidelines (note that a code of conduct is already present)

Community guidelines are available on the project website (<https://www.tesselle.org/contributing.html>) and in the repository (<https://github.com/tesselle/aion/blob/main/.github/CONTRIBUTING.md>).

> Additional suggestion: it might be interesting to add examples (with code) of how _aion_ can make interoperable formats used by other packages.

That's a good point. However, it's going to take me some time to find a real use case. I've just opened an issue to track progress on that (https://github.com/tesselle/aion/issues/4).
