+++ DESCRIPTION
@@ -25,7 +25,7 @@ BugReports: https://github.com/tesselle/aion/issues
 Depends: 
     R (>= 3.3)
 Imports: 
-    arkhe (>= 1.5.0),
+    arkhe (>= 1.6.0),
     graphics,
     grDevices,
     methods
