+++ man/fixed_julian.Rd
@@ -60,8 +60,8 @@ Other fixed date tools:
 \code{\link{as_decimal}()},
 \code{\link{as_fixed}()},
 \code{\link{as_year}()},
-\code{\link{fixed_gregorian}},
 \code{\link{fixed}()},
+\code{\link{fixed_gregorian}},
 \code{\link{format}()},
 \code{\link{pretty}()}
 }
