+++ man/as_year.Rd
@@ -61,9 +61,9 @@ Other fixed date tools:
 \code{\link{as_date}()},
 \code{\link{as_decimal}()},
 \code{\link{as_fixed}()},
+\code{\link{fixed}()},
 \code{\link{fixed_gregorian}},
 \code{\link{fixed_julian}},
-\code{\link{fixed}()},
 \code{\link{format}()},
 \code{\link{pretty}()}
 }
