+++ man/gregorian.Rd
@@ -47,8 +47,8 @@ J()
 \code{\link[=calendar]{calendar()}}
 
 Other calendar tools: 
-\code{\link{calendar_get}},
 \code{\link{calendar}()},
+\code{\link{calendar_get}},
 \code{\link{convert}()},
 \code{\link{is}()},
 \code{\link{julian}()}
