+++ man/fixed_gregorian.Rd
@@ -94,8 +94,8 @@ Other fixed date tools:
 \code{\link{as_decimal}()},
 \code{\link{as_fixed}()},
 \code{\link{as_year}()},
-\code{\link{fixed_julian}},
 \code{\link{fixed}()},
+\code{\link{fixed_julian}},
 \code{\link{format}()},
 \code{\link{pretty}()}
 }
