+++ man/is.Rd
@@ -31,8 +31,8 @@ Test inheritance relationships between an object and a calendar class.
 }
 \seealso{
 Other calendar tools: 
-\code{\link{calendar_get}},
 \code{\link{calendar}()},
+\code{\link{calendar_get}},
 \code{\link{convert}()},
 \code{\link{gregorian}},
 \code{\link{julian}()}
