### 2024-03-14 #6210 @nfrerebeau
@martinfleis @steko I have just updated the README and vignette manual to use a real-world example as suggested. See https://github.com/tesselle/aion/commit/8c2450a7dda33f0b0badbde261b8a8aa5b985aeb

@martinfleis @sebastien-plutniak Regarding https://github.com/tesselle/aion/issues/4, this will require writing an additional vignette. If it's not absolutely necessary for this review, it's on my roadmap for later.
