### 2024-03-28 #6210 @martinfleis
Hi @sebastien-plutniak,
in light of https://github.com/openjournals/joss-reviews/issues/6210#issuecomment-1963758055, are there any other outstanding issues preventing the publication of the paper?
