### 2024-04-04 #6210 @martinfleis
Thank you @sebastien-plutniak and @steko!

@nfrerebeau please refer to the post-review tasks listed in https://github.com/openjournals/joss-reviews/issues/6210#issuecomment-2036413879 and let me know once you're done with them. Let me know if you have any questions.
