+++ man/aion-package.Rd
@@ -14,7 +14,7 @@ A toolkit for archaeological time series. This package provides a system of clas
 \tabular{ll}{
 \strong{Package:} \tab aion \cr
 \strong{Type:} \tab Package \cr
-\strong{Version:} \tab 1.0.2 \cr
+\strong{Version:} \tab 1.0.3 \cr
 \strong{License:} \tab GPL-3 \cr
 \strong{Zenodo:} \tab \doi{10.5281/zenodo.8032278} \cr
 }
