+++ cran-comments.md
@@ -1,5 +1,5 @@
 ## Test environments
-* Local R installation: R 4.3.1
+* Local R installation: R 4.3.3
 * Ubuntu (on GitHub Actions): R devel, release, oldrel
 * Windows (on GitHub Actions and win-builder): R devel, release, oldrel
 * MacOS (on GitHub Actions): R release
