+++ README.md
@@ -56,20 +56,20 @@ offers a simple API that can be used by other specialized packages.
 
     To cite aion in publications use:
 
-      Frerebeau N, Roe J (2023). _aion: Archaeological Time Series_.
+      Frerebeau N, Roe J (2024). _aion: Archaeological Time Series_.
       Université Bordeaux Montaigne, Pessac, France.
       doi:10.5281/zenodo.8032278 <https://doi.org/10.5281/zenodo.8032278>,
-      R package version 1.0.2, <https://packages.tesselle.org/aion/>.
+      R package version 1.0.3, <https://packages.tesselle.org/aion/>.
 
     A BibTeX entry for LaTeX users is
 
       @Manual{,
         author = {Nicolas Frerebeau and Joe Roe},
         title = {{aion: Archaeological Time Series}},
-        year = {2023},
+        year = {2024},
         organization = {Université Bordeaux Montaigne},
         address = {Pessac, France},
-        note = {R package version 1.0.2},
+        note = {R package version 1.0.3},
         url = {https://packages.tesselle.org/aion/},
         doi = {10.5281/zenodo.8032278},
       }
