+++ revdep/README.md
@@ -2,24 +2,24 @@
 
 |field    |value                                                                        |
 |:--------|:----------------------------------------------------------------------------|
-|version  |R version 4.3.1 (2023-06-16)                                                 |
-|os       |Ubuntu 22.04.3 LTS                                                           |
+|version  |R version 4.3.3 (2024-02-29)                                                 |
+|os       |Ubuntu 22.04.4 LTS                                                           |
 |system   |x86_64, linux-gnu                                                            |
 |ui       |RStudio                                                                      |
 |language |(EN)                                                                         |
 |collate  |fr_FR.UTF-8                                                                  |
 |ctype    |fr_FR.UTF-8                                                                  |
 |tz       |Europe/Paris                                                                 |
-|date     |2023-10-26                                                                   |
-|rstudio  |2023.09.0+463 Desert Sunflower (desktop)                                     |
+|date     |2024-04-04                                                                   |
+|rstudio  |2023.12.1+402 Ocean Storm (desktop)                                          |
 |pandoc   |3.1.1 @ /usr/lib/rstudio/resources/app/bin/quarto/bin/tools/ (via rmarkdown) |
 
 # Dependencies
 
-|package |old   |new        |Δ  |
-|:-------|:-----|:----------|:--|
-|aion    |1.0.1 |1.0.1.9000 |*  |
-|arkhe   |1.3.0 |1.3.0      |   |
+|package |old   |new   |Δ  |
+|:-------|:-----|:-----|:--|
+|aion    |1.0.2 |1.0.3 |*  |
+|arkhe   |1.6.0 |1.6.0 |   |
 
 # Revdeps
 
