+++ CITATION.cff
@@ -1,5 +1,5 @@
 # -----------------------------------------------------------
-# CITATION file created with {cffr} R package, v0.5.0
+# CITATION file created with {cffr} R package, v1.0.0
 # See also: https://docs.ropensci.org/cffr/
 # -----------------------------------------------------------
  
@@ -8,13 +8,13 @@ message: 'To cite package "aion" in publications use:'
 type: software
 license: GPL-3.0-or-later
 title: 'aion: Archaeological Time Series'
-version: 1.0.2
+version: 1.0.3
 doi: 10.5281/zenodo.8032278
 abstract: A toolkit for archaeological time series. This package provides a system
   of classes and methods to represent and work with archaeological time series. Dates
   are represented as "rata die" and can be converted to (virtually) any calendar defined
-  by Reingold and Dershowitz (2018) <doi:10.1017/9781107415058>. This packages offers
-  a simple API that can be used by other specialized packages.
+  by Reingold and Dershowitz (2018) <https://doi.org/10.1017/9781107415058>. This
+  packages offers a simple API that can be used by other specialized packages.
 authors:
 - family-names: Frerebeau
   given-names: Nicolas
@@ -40,14 +40,13 @@ preferred-citation:
     email: joseph.roe@iaw.unibe.ch
     orcid: https://orcid.org/0000-0002-1011-1244
     affiliation: Universität Bern
-  year: '2023'
-  location:
-    name: Pessac, France
-  notes: R package version 1.0.2
-  url: https://packages.tesselle.org/aion/
-  doi: 10.5281/zenodo.8032278
+  year: '2024'
   institution:
     name: Université Bordeaux Montaigne
+    address: Pessac, France
+  notes: R package version 1.0.3
+  url: https://packages.tesselle.org/aion/
+  doi: 10.5281/zenodo.8032278
 repository: https://CRAN.R-project.org/package=aion
 repository-code: https://github.com/tesselle/aion
 url: https://packages.tesselle.org/aion/
@@ -70,11 +69,10 @@ references:
   url: https://www.R-project.org/
   authors:
   - name: R Core Team
-  location:
-    name: Vienna, Austria
-  year: '2023'
   institution:
     name: R Foundation for Statistical Computing
+    address: Vienna, Austria
+  year: '2024'
   version: '>= 3.3'
 - type: software
   title: arkhe
@@ -88,41 +86,52 @@ references:
     email: nicolas.frerebeau@u-bordeaux-montaigne.fr
     orcid: https://orcid.org/0000-0001-5759-4944
     affiliation: Université Bordeaux Montaigne
-  year: '2023'
-  version: '>= 1.3.0'
+  year: '2024'
+  version: '>= 1.6.0'
 - type: software
   title: graphics
   abstract: 'R: A Language and Environment for Statistical Computing'
   notes: Imports
   authors:
   - name: R Core Team
-  location:
-    name: Vienna, Austria
-  year: '2023'
   institution:
     name: R Foundation for Statistical Computing
+    address: Vienna, Austria
+  year: '2024'
 - type: software
   title: grDevices
   abstract: 'R: A Language and Environment for Statistical Computing'
   notes: Imports
   authors:
   - name: R Core Team
-  location:
-    name: Vienna, Austria
-  year: '2023'
   institution:
     name: R Foundation for Statistical Computing
+    address: Vienna, Austria
+  year: '2024'
 - type: software
   title: methods
   abstract: 'R: A Language and Environment for Statistical Computing'
   notes: Imports
[…]
