### 2024-04-04 #6210 @nfrerebeau
@martinfleis I have corrected my affiliation in the article (https://github.com/tesselle/aion/commit/362f40a3).
