### 2024-04-04 #6210 @nfrerebeau
Release: [v1.0.3](https://github.com/tesselle/aion/releases/tag/v1.0.3)
DOI : [10.5281/zenodo.10926190](https://doi.org/10.5281/zenodo.10926190)
