### 2024-04-04 #6210 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
