### 2024-04-07 #6210 @oliviaguest
@nfrerebeau can you amend the double parentheses please? So things like ```(e.g., Bchron by Haslett & Parnell (2008) or rcarbon by Crema & Bevan (2021))``` become something like ```(e.g., Bchron by Haslett & Parnell, 2008; or rcarbon by Crema & Bevan, 2021)```. If you don't know how to do it: remove the square brackets. ☺️
