### 2024-04-08 ca67754
Remove double parentheses
