### 2024-04-11 #6210 @arfon
Yeah, I'm afraid this (linking to ORCIDs for reviewers) isn't currently supported sorry.
