+++ NEWS.md
@@ -1,5 +1,5 @@
 # aion 1.0.3
-* Published in the *Journal of Open Source Software*.
+* Published in the [*Journal of Open Source Software*](https://doi.org/10.21105/joss.06210).
 
 # aion 1.0.2
 ## Bugfixes & changes
