+++ NEWS.md
@@ -1,3 +1,5 @@
+# aion 1.0.3.9000
+
 # aion 1.0.3
 * Published in the [*Journal of Open Source Software*](https://doi.org/10.21105/joss.06210).
 
