+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: aion
 Title: Archaeological Time Series
-Version: 1.0.3
+Version: 1.0.3.9000
 Authors@R: c(
     person("Nicolas", "Frerebeau", , "nicolas.frerebeau@u-bordeaux-montaigne.fr", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0001-5759-4944", affiliation = "Université Bordeaux Montaigne")),
